refactor: make Metric type flexible
